requirements.txt
Django==4.2
psycopg2==2.9.2
django-autoslug==1.9.8
django-countries==7.2.1
Pillow==8.3.2
phonenumbers==8.12.33
psycopg2-binary==2.9.2
flake8==3.9.2
black==22.1.0
isort==5.9.3
PyJWT==2.1.0
django-celery-beat==2.1.0
django-countries==7.2.1
django-crispy-forms==1.13.0
django-decouple==2.1
sqlparse==0.4.2
bootstrap4==0.1.0
django-filter==21.1
django-allauth==0.51.0
dj-static==0.0.6
dj-database-url==1.0.0
whitenoise==6.2.0
gunicorn==20.1.0
rest-social-auth==5.0.0
django-environ==0.8.0
xlwt==1.3.0
django-import-export==2.8.0
djangorestframework==3.13.1
qrcode==7.4.2
